requirements.txt
streamlit
torch
torchvision
Pillow
numpy
opencv-python
plotly
tensorflow-cpu
matplotlib
opencv-python-headless

